+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Use enum names in generated test functions when defined in `burstValues`.
 * Make the test class constructor with default arguments synthetic. Fixes JUnit5 compatibility.
+* Support AGP's built-in Kotlin.
 
 
 ## [2.9.0] *(2025-09-04)*